Refactor withdrawal and deposit commands to use UTXO outpoints instead of txid and vout

+++ src/backend.rs
@@ -272,7 +272,7 @@ pub(crate) async fn send_withdrawal_signatures_to_operators(
     withdrawal_index: u32,
     signature: &str,
     config: &BridgeCliConfig,
-    amount: u64,
+    amount: f64,
 ) -> Result<(), BridgeCliError> {
     let url = config
         .citrea_backend_endpoint // As long as URL is only base, no trailing / (slash) is needed
